Clean-up import cruft and test package namespaces.
This commit removes a number of crufty imports from some of the support
library.  Secondarily, Go Version 1 seems to expect that tests for a
given package coexist in the same directory as the package under test
along with having the same package name.

Also, this change builds on https://github.com/pomack/thrift4go/pull/20,
which explains the initial size of it.  Once the formatting rules are
reconciled, this will be much smaller.